# swagger.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Voca Commerce Server",
        "description": "API for managing commerce interaction",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://voca-commerce-server.vercel.app/"
        },
        {
            "url": "http://localhost:3000"
        }
    ],
    "tags": [
        {
            "name": "admin register"
        },
        {
            "name": "register"
        },
        {
            "name": "login"
        },
        {
            "name": "admin product"
        },
        {
            "name": "product"
        }
    ],
    "paths": {
        "/v1/admin/register": {
            "post": {
                "tags": [
                    "admin register"
                ],
                "summary": "Register as admin",
                "operationId": "registerAdmin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/register"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/register": {
            "post": {
                "tags": [
                    "register"
                ],
                "summary": "Register as user",
                "operationId": "registerUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/register"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "login"
                ],
                "summary": "Login into server",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/login"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/products": {
            "get": {
                "tags": [
                    "product"
                ],
                "summary": "Get products by filters",
                "operationId": "getProductsByFilter",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Product name to filter",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Field to sort",
                        "schema": {
                            "type": "string",
                            "default": "id",
                            "enum": [
                                "id",
                                "createdAt",
                                "updatedAt",
                                "price"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Sort the result in ascending or descending order",
                        "schema": {
                            "type": "string",
                            "default": "desc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "admin product"
                ],
                "summary": "Insert a new product",
                "operationId": "createProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/createProduct"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/products/{id}": {
            "get": {
                "tags": [
                    "admin product"
                ],
                "summary": "Get a detailed product",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "patch": {
                "tags": [
                    "admin product"
                ],
                "summary": "Partial edit product by id",
                "operationId": "updateProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/updateProduct"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin product"
                ],
                "summary": "Remove product by id",
                "operationId": "deleteProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/products/{id}/soft-delete": {
            "patch": {
                "tags": [
                    "admin product"
                ],
                "summary": "Soft delete product by id",
                "operationId": "softDeleteProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/v1/products/{id}/restore": {
            "patch": {
                "tags": [
                    "admin product"
                ],
                "summary": "Restore product by id from soft deletion",
                "operationId": "restoreProductById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the product",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "register": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "securepassword"
                    }
                }
            },
            "login": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "admin@gmail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "adminpassword"
                    }
                }
            },
            "updateProduct": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "My Product"
                    },
                    "description": {
                        "type": "string",
                        "example": "Very good product with affordable price"
                    },
                    "price": {
                        "type": "number",
                        "example": 549321
                    },
                    "available": {
                        "type": "boolean",
                        "example": false
                    }
                }
            },
            "createProduct": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "My New Product"
                    },
                    "description": {
                        "type": "string",
                        "example": "State of the art product with affordable price"
                    },
                    "price": {
                        "type": "number",
                        "example": 123987
                    },
                    "available": {
                        "type": "boolean",
                        "example": true
                    }
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}